esp32의 남은 시간, 자리 이동 대응 로직 추가
- LinkedBlockingQueue의 경우 take는 무한 대기, poll은 지정된 시간 대기 후 null 반환
- esp32의 값 받는 로직 수정에 따른 prelong 로직 수정
- esp32의 남은 시간 가져오는 로직 추가
- 각 controller 메서드에 return 추가

+++ src/main/kotlin/com/example/payments/dto/esp32/ESP32PreLongRequestDto.kt
@@ -1,7 +1,7 @@
 package com.example.payments.dto.esp32
 
 data class ESP32PreLongRequestDto(
-    val time: Int = 0
+    val time: Long = 0
 ) {
     init {
         require(time >= -1) { "시간 값은 -1 이상의 정수형이어야 합니다." }
